+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 - Fixes behavior for setting environment variables in a `config.toml` as relative paths to properly result in an absolute path one directory above the definition of the `config.toml`
 - Fixes `orbit env` to display the correct values for environment variables (including those found in `config.toml` files under the `[env]` table)
 - Improves error messages around the `orbit doc` command when trying to open the generated documentation using `--open`
+- Correctly generates documentation for projects that have nested relative dependencies
 
 ## 0.28.1